+++ source/ecs/components/Networked.yaml
@@ -1,0 +1,25 @@
+
+
+# Useful for sending or receiving data associated with an entity over the network
+Networked:
+
+  networkID: [final<int>, rand()]
+  templateHash: final<int>
+
+  _hash:
+    [networkID]
+
+  _cpp_only:
+
+    # lists of data to send or receive:
+    toSend: NetworkedDataList
+    toReceive: NetworkedDataList
+
+    # List of data that will be:
+    #  - send if Entity has LocalPlayer-component
+    #  - received if Entity does not have a LocalPlayer-component.
+
+    sendIfLocalPlayerReceiveOtherwise: NetworkedDataList
+
+
+    dataPresence: std::map<size_t, bool>
