+++ source/ecs/systems/PlatformerMovementSystem.h
@@ -6,8 +6,8 @@
 #include <input/key_input.h>
 #include "EntitySystem.h"
 #include "../../level/Level.h"
-#include "../components/physics/Physics.h"
-#include "../components/graphics/AsepriteView.h"
+#include "../../generated/Physics.hpp"
+#include "../../generated/AsepriteView.hpp"
 #include "../../generated/PlayerControlled.hpp"
 #include "../../generated/Aiming.hpp"
 #include "../../generated/PlatformerMovement.hpp"
