+++ source/ecs/systems/SpawningSystem.cpp
@@ -1,6 +1,6 @@
 
 #include "SpawningSystem.h"
-#include "../components/physics/Physics.h"
+#include "../../generated/Physics.hpp"
 #include "../../generated/Spawning.hpp"
 
 void SpawningSystem::update(double deltaTime, EntityEngine *room)
