+++ source/ecs/systems/networking/NetworkingSystem.h
@@ -4,7 +4,6 @@
 
 
 #include "../EntitySystem.h"
-#include "../../components/Networked.h"
 #include "../../../game/session/Player.h"
 #include "../../../game/session/multiplayer/packets.h"
