+++ source/game/session/multiplayer/MultiplayerServerSession.cpp
@@ -1,8 +1,7 @@
 
 #include <gu/profiler.h>
 #include "MultiplayerServerSession.h"
-#include "../../../ecs/components/physics/Physics.h"
-#include "../../../ecs/components/Networked.h"
+#include "../../../generated/Physics.hpp"
 
 using namespace Packet;
 using namespace Packet::from_player;
