+++ source/macro_magic/includes_for_generated_structs.h
@@ -2,3 +2,11 @@
 #include "component.h"
 #include <audio/audio.h>
 #include "../ecs/entity_templates/LuaEntityTemplate.h"
+
+#include "../ecs/systems/networking/NetworkedDataList.h"
+#include "../ecs/systems/networking/NetworkedDataGroup.h"
+
+#include <set>
+
+struct AABB;
+#include "../ecs/systems/physics/TerrainCollisionDetector.h"
