+++ source/rendering/level/room/RoomScreen.h
@@ -14,8 +14,7 @@
 #include "../../../level/room/Room.h"
 #include "RoomEditor.h"
 #include "DebugTileRenderer.h"
-#include "../../../ecs/components/physics/Physics.h"
-#include "../../../ecs/components/Networked.h"
+#include "../../../generated/Physics.hpp"
 #include "../../../ecs/EntityInspector.h"
 #include "CameraMovement.h"
 #include "tile_map/TileMapRenderer.h"
@@ -27,7 +26,6 @@
 #include "../../sprites/MegaSpriteSheet.h"
 #include "../../sprites/SpriteRenderer.h"
 #include "../../PolylineRenderer.h"
-#include "../../../ecs/components/body_parts/Leg.h"
 #include "../../../ecs/components/Polyline.h"
 #include "../../../ecs/components/physics/PolyPlatform.h"
 #include "../../../ecs/entity_templates/LuaEntityTemplate.h"
