+++ source/rendering/level/room/blood_splatter/BloodSplatterRenderer.cpp
@@ -1,10 +1,11 @@
 
 #include <graphics/orthographic_camera.h>
+#include <graphics/3d/vert_buffer.h>
 #include "BloodSplatterRenderer.h"
 
 #include "../../../../level/room/Room.h"
 #include "../../../../ecs/components/graphics/BloodDrop.h"
-#include "../../../../ecs/components/physics/Physics.h"
+#include "../../../../generated/Physics.hpp"
 #include "../../../../generated/Spawning.hpp"
